add a request cache to all popular endpoints

+++ back-end/src/api/lobbies/leave-game.ts
@@ -1,7 +1,6 @@
 import { Request, Response } from 'express';
 import { Game } from '../../models/Game.model';
 import { JoinGameRequest } from '../../models/JoinGameRequest.model';
-import { getUser } from '../../services/osu-api';
 
 export async function leaveGame(req: Request, res: Response) {
   const { requestId } = req.body;
@@ -12,15 +11,15 @@ export async function leaveGame(req: Request, res: Response) {
     return res.status(404).end();
   }
 
+  if (!verifyRequest.verified) {
+    return res.status(200).end();
+  }
+
   const game = await Game.findById(req.params.id);
 
   if (!game) {
     return res.status(404).end();
   }
-
-  const player = game.players.find(
-    p => p.username.toLowerCase() === verifyRequest.username.toLowerCase(),
-  );
 
   if (verifyRequest.verified) {
     await JoinGameRequest.deleteOne({ _id: verifyRequest.id });
